kafka: split scram credential into separate header
Intended to avoid some circular dependencies between the generic
credential store and the specific handlers for sasl mechanisms.
