+++ src/v/kafka/security/scram_algorithm.h
@@ -13,6 +13,7 @@
 #include "hashing/secure.h"
 #include "kafka/server/logger.h"
 #include "random/generators.h"
+#include "kafka/security/scram_credential.h"
 
 #include <absl/container/node_hash_map.h>
 
@@ -43,31 +44,6 @@ public:
 
 private:
     ss::sstring _msg;
-};
-
-class scram_credential {
-public:
-    scram_credential() noexcept = default;
-
-    scram_credential(
-      bytes salt, bytes server_key, bytes stored_key, int iterations) noexcept
-      : _salt(std::move(salt))
-      , _server_key(std::move(server_key))
-      , _stored_key(std::move(stored_key))
-      , _iterations(iterations) {}
-
-    const bytes& salt() const { return _salt; }
-    const bytes& server_key() const { return _server_key; }
-    const bytes& stored_key() const { return _stored_key; }
-    int iterations() const { return _iterations; }
-
-private:
-    friend std::ostream& operator<<(std::ostream&, const scram_credential&);
-
-    bytes _salt;
-    bytes _server_key;
-    bytes _stored_key;
-    int _iterations{0};
 };
 
 /**
@@ -185,7 +161,6 @@ private:
     bytes _signature;
 };
 
-std::ostream& operator<<(std::ostream&, const scram_credential&);
 std::ostream& operator<<(std::ostream&, const client_first_message&);
 std::ostream& operator<<(std::ostream&, const server_first_message&);
 std::ostream& operator<<(std::ostream&, const client_final_message&);
